Fix many -Wconversion compiler warnings

+++ unix/tkUnixEvent.c
@@ -184,13 +184,13 @@ TkpOpenDisplay(
     if (WidthMMOfScreen(DefaultScreenOfDisplay(display)) <= 0) {
 	int mm;
 
-	mm = WidthOfScreen(DefaultScreenOfDisplay(display)) * (25.4 / 75.0);
+	mm = (int)(WidthOfScreen(DefaultScreenOfDisplay(display)) * (25.4 / 75.0));
 	WidthMMOfScreen(DefaultScreenOfDisplay(display)) = mm;
     }
     if (HeightMMOfScreen(DefaultScreenOfDisplay(display)) <= 0) {
 	int mm;
 
-	mm = HeightOfScreen(DefaultScreenOfDisplay(display)) * (25.4 / 75.0);
+	mm = (int)(HeightOfScreen(DefaultScreenOfDisplay(display)) * (25.4 / 75.0));
 	HeightMMOfScreen(DefaultScreenOfDisplay(display)) = mm;
     }
 
@@ -592,8 +592,8 @@ TkUnixDoOneXEvent(
 	    blockTime.tv_usec = 0;
 	}
 	fd = ConnectionNumber(dispPtr->display);
-	index = fd/(NBBY*sizeof(fd_mask));
-	bit = ((fd_mask)1) << (fd%(NBBY*sizeof(fd_mask)));
+	index = fd/(NBBY*(int)sizeof(fd_mask));
+	bit = ((fd_mask)1) << (fd%(NBBY*(int)sizeof(fd_mask)));
 	readMask[index] |= bit;
 	if (numFdBits <= fd) {
 	    numFdBits = fd+1;
@@ -618,8 +618,8 @@ TkUnixDoOneXEvent(
     for (dispPtr = TkGetDisplayList(); dispPtr != NULL;
 	    dispPtr = dispPtr->nextPtr) {
 	fd = ConnectionNumber(dispPtr->display);
-	index = fd/(NBBY*sizeof(fd_mask));
-	bit = ((fd_mask)1) << (fd%(NBBY*sizeof(fd_mask)));
+	index = fd/(NBBY*(int)sizeof(fd_mask));
+	bit = ((fd_mask)1) << (fd%(NBBY*(int)sizeof(fd_mask)));
 	if ((readMask[index] & bit) || (QLength(dispPtr->display) > 0)) {
 	    DisplayFileProc(dispPtr, TCL_READABLE);
 	}
